added figures to pipelines in examples/pipelines/tests
