+++ examples/pipelines/tests/join_join.js
@@ -8,19 +8,17 @@ const {
   fork
 } = require('../../..')
 
-const task0 = task({name: 'coco'}, () => `echo "something0"`)
+const task0 = task({name: 'task0'}, () => `echo "something0"`)
 
-const task1 = task({name: 'nut'}, () => `echo "something1"`)
+const task2 = task({name: 'task2'}, () => `echo "something2"`)
 
-const task2 = task({name: 'foo'}, () => `echo "something2"`)
+const task3 = task({name: 'task3'}, () => `echo "something3"`)
 
-const task3 = task({name: 'bar'}, () => `echo "something3"`)
+const task4 = task({name: 'task4'}, () => `echo "something4"`)
 
-const task4 = task({name: 'test'}, () => `echo "something4"`)
+const task5 = task({name: 'task5'}, () => `echo "something5"`)
 
-const task5 = task({name: 'test1'}, () => `echo "something5"`)
-
-const task6 = task({name: 'test6'}, () => `echo "something6"`)
+const task6 = task({name: 'task6'}, () => `echo "something6"`)
 
 const pipeline = join(
   task0,
