+++ examples/pipelines/tests/simple_fork.js
@@ -16,9 +16,6 @@ const task2 = task({name: 'foo'}, () => `echo "something2"`)
 
 const task3 = task({name: 'bar'}, () => `echo "something3"`)
 
-const task4 = task({name: 'test'}, () => `echo "something4"`)
-
-
-const pipeline = join(task0, fork(task1, task2), task3, task4)
+const pipeline = join(task0, fork(task1, task2), task3)
 
 pipeline()
